RobotTest: PcInterface fixups, move LwipUdpInterface into own files
