+++ RobotTest/src/PcInterface.cpp
@@ -1,133 +1,4 @@
 #include <PcInterface.h>
-#include <UdpInterface.h>
-
-// LwipUdpInterface is the real i.e. hardware-facing implementation of the
-// abstract UdpInterface class. This will call out to LwIP and any
-// lower-level networking functions to implement UDP functionality.
-// For now, it is never used outside of PcInterface, so the class
-// definition and implementation live in PcInterface.cpp.
-// Functions have no parameters, and instead pass the state information
-// they were initialized with (members of this class). This means the PcInterface,
-// UdpInterface, and mocking classes do not have to care about how the
-// networking functions are implemented under the hood - only the methods
-// called.
-// Another approach rather than having stateful members is to pass a
-// void* to every interface function, and pass data as needed into there.
-class LwipUdpInterface : public UdpInterface {
-public:
-	LwipUdpInterface();
-	// TODO: parameterized constructor.
-	~LwipUdpInterface();
-
-	// Abstract UdpInterface class overrides.
-	bool udpNew();
-	bool udpBind();
-	bool udpRecv();
-	bool udpRemove();
-	bool ethernetifInput();
-	bool udpConnect();
-	bool udpSend();
-	bool udpDisconnect();
-	bool pbufFreeRx();
-	bool pbufFreeTx();
-	bool waitRecv();
-	bool packetToBytes(uint8_t *_byteArray);
-	bool bytesToPacket(const uint8_t *_byteArray);
-
-	// Helpers
-	void setRecvCallbackPbuf(void* _recvCallbackPbuf);
-private:
-	// FIXME: These need to be of the right types for the networking stack.
-	const int *pcb = nullptr;
-	const int *ipaddr = nullptr;
-	const int *ipaddrPc = nullptr;
-	const int port = 0;
-	const int portPc = 0;
-	const int *netif = nullptr;
-	const int *recvSemaphore = nullptr;
-
-	// Set by recvCallback.
-	void *recvCallbackPbuf = nullptr;
-
-	int *txPbuf = nullptr;
-	// TODO: synchronize access to pbufs
-};
-
-// FIXME: use correct networking stack types in callback function.
-void recvCallback(LwipUdpInterface *arg, void *pcb, void *p,
-		const void *addr, int port);
-
-LwipUdpInterface::LwipUdpInterface() {
-
-}
-
-LwipUdpInterface::~LwipUdpInterface() {
-
-}
-
-bool LwipUdpInterface::udpNew() {
-	return true;
-}
-
-bool LwipUdpInterface::udpBind() {
-	return true;
-}
-
-bool LwipUdpInterface::udpRecv() {
-	return true;
-}
-
-bool LwipUdpInterface::udpRemove() {
-	return true;
-}
-
-bool LwipUdpInterface::ethernetifInput() {
-	return true;
-}
-
-bool LwipUdpInterface::udpConnect() {
-	return true;
-}
-
-bool LwipUdpInterface::udpSend() {
-	return true;
-}
-
-bool LwipUdpInterface::udpDisconnect() {
-	return true;
-}
-
-bool LwipUdpInterface::pbufFreeRx() {
-	return true;
-}
-
-bool LwipUdpInterface::pbufFreeTx() {
-	return true;
-}
-
-bool LwipUdpInterface::waitRecv() {
-	return true;
-}
-
-bool LwipUdpInterface::packetToBytes(uint8_t *_byteArray) {
-	return true;
-}
-
-bool LwipUdpInterface::bytesToPacket(const uint8_t *_byteArray) {
-	return true;
-}
-
-void LwipUdpInterface::setRecvCallbackPbuf(void* _recvCallbackPbuf) {
-	recvCallbackPbuf = _recvCallbackPbuf;
-}
-
-void recvCallback(LwipUdpInterface *arg, void *pcb, void *p,
-		const void *addr, int port) {
-	// TODO: synchronize access to these members (stop them getting clobbered before read by PcInterface)
-	arg->setRecvCallbackPbuf(p);
-	// TODO: release recvSemaphore
-}
-
 
 PcInterface::PcInterface() {
 	// No need to initialize any members here; this is done in PcInterface.h
@@ -199,7 +70,6 @@ bool PcInterface::transmit() {
 		udpInterface->udpSend();
 		udpInterface->udpDisconnect();
 		udpInterface->pbufFreeTx();
-		// Should not return.
 		return true;
 	default:
 		return false;
